+++ src/UploadClient.ts
@@ -4,20 +4,20 @@ import fileFrom from './fileFrom/fileFrom'
 import groupFrom from './groupFrom/groupFrom'
 
 /* Types */
-import {FileFromEnum, FileUploadInterface} from './fileFrom/types'
 import {FileData, SettingsInterface} from './types'
+import {FileUploadInterface} from './fileFrom/types'
 import {Url} from './api/fromUrl'
 import {UploadAPIInterface, Uuid} from './api/types'
-import {GroupFromEnum, GroupUploadInterface} from './groupFrom/types'
+import {GroupUploadInterface} from './groupFrom/types'
 
 export interface UploadClientInterface {
   updateSettings(newSettings: SettingsInterface): void
 
   getSettings(): SettingsInterface
 
-  fileFrom(from: FileFromEnum, data: FileData | Url | Uuid, settings?: SettingsInterface): FileUploadInterface
+  fileFrom(data: FileData | Url | Uuid, settings?: SettingsInterface): FileUploadInterface
 
-  groupFrom(from: GroupFromEnum, data: FileData[] | Url[] | Uuid[], settings?: SettingsInterface): GroupUploadInterface
+  groupFrom(data: FileData[] | Url[] | Uuid[], settings?: SettingsInterface): GroupUploadInterface
 }
 
 class UploadClient implements UploadClientInterface {
@@ -45,15 +45,15 @@ class UploadClient implements UploadClientInterface {
     return this.settings
   }
 
-  fileFrom(from: FileFromEnum, data: FileData | Url | Uuid, settings: SettingsInterface = {}): FileUploadInterface {
-    return fileFrom(from, data, {
+  fileFrom(data: FileData | Url | Uuid, settings: SettingsInterface = {}): FileUploadInterface {
+    return fileFrom(data, {
       ...this.settings,
       ...settings,
     })
   }
 
-  groupFrom(from: GroupFromEnum, data: FileData[] | Url[] | Uuid[], settings: SettingsInterface = {}): GroupUploadInterface {
-    return groupFrom(from, data, {
+  groupFrom(data: FileData[] | Url[] | Uuid[], settings: SettingsInterface = {}): GroupUploadInterface {
+    return groupFrom(data, {
       ...this.settings,
       ...settings,
     })