+++ src/index.ts
@@ -7,6 +7,5 @@ export {Uuid, UploadAPIInterface} from './api/types'
 export {FileUploadInterface} from './fileFrom/types'
 export {UploadClientInterface} from './UploadClient'
 export {RequestOptionsInterface, RequestInterface} from './api/request/types'
-export {FileFromEnum} from './fileFrom/types'
 
 export default UploadClient