+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * `addUpdateSettingsListener` and `removeUpdateSettingsListener` from 
 `UploadClientInterface`.
+* Removed `from` param from `fileFrom` and `groupFrom`. 
+* Removed `FileFromEnum` and `GroupFromEnum`.
 
 ## [1.0.0-alpha.5]